+++ app/Http/Controllers/Api/TagController.php
@@ -412,6 +412,7 @@ class TagController extends ApiController
      * @param array criteria[dataset_criteria][category_ids] - optional
      * @param array criteria[dataset_criteria][tag_ids] - optional
      * @param array criteria[dataset_criteria][formats] - optional
+     * @param array criteria[dataset_criteria][access] - optional
      * @param array criteria[dataset_criteria][terms_of_use_ids] - optional
      * @param boolean criteria[dataset_criteria][reported] - optional
      * @param array criteria[dataset_ids] - optional
@@ -456,6 +457,7 @@ class TagController extends ApiController
                 'terms_of_use_ids.*'  => 'int|digits_between:1,10|exists:terms_of_use,id',
                 'formats'             => 'nullable|array|min:1',
                 'formats.*'           => 'string|in:'. implode(',', Resource::getFormats()),
+                'access'              => 'nullable|int|in:'. implode(',', array_keys(DataSet::getAccessTypes())),
                 'reported'            => 'nullable|boolean',
             ]);
         }
@@ -489,6 +491,9 @@ class TagController extends ApiController
                         $q->whereHas('DataSetTags', function($qr) use ($dsCriteria) {
                             $qr->whereIn('tag_id', $dsCriteria['tag_ids']);
                         });
+                    }
+                    if (!empty($dsCriteria['access'])) {
+                        $q->where('data_sets.access', $dsCriteria['access']);
                     }
                     if (!empty($dsCriteria['terms_of_use_ids'])) {
                         $q->whereIn('terms_of_use_id', $dsCriteria['terms_of_use_ids']);