+++ app/Http/Controllers/Api/TermsOfUseController.php
@@ -366,6 +366,7 @@ class TermsOfUseController extends ApiController
      * @param array criteria[dataset_criteria][category_ids] - optional
      * @param array criteria[dataset_criteria][tag_ids] - optional
      * @param array criteria[dataset_criteria][formats] - optional
+     * @param array criteria[dataset_criteria][access] - optional
      * @param array criteria[dataset_criteria][terms_of_use_ids] - optional
      * @param boolean criteria[dataset_criteria][reported] - optional
      * @param array criteria[dataset_ids] - optional
@@ -411,6 +412,7 @@ class TermsOfUseController extends ApiController
                 'terms_of_use_ids.*'  => 'int|digits_between:1,10|exists:terms_of_use,id',
                 'formats'             => 'nullable|array|min:1',
                 'formats.*'           => 'string|in:'. implode(',', Resource::getFormats()),
+                'access'              => 'nullable|int|in:'. implode(',', array_keys(DataSet::getAccessTypes())),
                 'reported'            => 'nullable|boolean',
             ]);
         }
@@ -455,6 +457,9 @@ class TermsOfUseController extends ApiController
                         'data_sets.id',
                         DB::table('data_set_tags')->select('data_set_id')->distinct()->whereIn('tag_id', $dsCriteria['tag_ids'])
                     );
+                }
+                if (!empty($dsCriteria['access'])) {
+                    $data->where('data_sets.access', $dsCriteria['access']);
                 }
                 if (!empty($dsCriteria['terms_of_use_ids'])) {
                     $data->whereIn('terms_of_use_id', $dsCriteria['terms_of_use_ids']);